Handled app shutdown during render.Pretty janky solution that leaves the worker threads running until they finish the tile they're working on. Maybe put warning message about this for when user force closes app during render.

+++ Source/engine/renderer/renderer.cpp
@@ -284,7 +284,7 @@ static b32 render_tile_from_camera(RenderInfo& info, RNG_Stream* rng_stream)
 	{
 		rt = &info.twq.jobs[(int32)tile_no - 1];
 	}
-	ATP_BLOCK_M(Tiles, tile_no-1);
+	ATP_BLOCK_M(Tiles, (uint32)tile_no-1);
 
 
 	tile_ = &rt->tile;
